data_access_byte_ wrong size
I've been trying to compile this code with avrstudio, the display has been showing 128 for all the variables. 

After much time debugging I have established that the storage data_access_byte_ is only set to one element! So any writes and reads were out of bounds. Is kNumParameters the right ? I added 1 as you seem to do that to the index in the code.